refactor: move DeleteButton into buttons folder
