+++ dev-client/src/screens/SiteSettingsScreen/SiteSettingsScreen.tsx
@@ -22,7 +22,7 @@ import {Fab} from 'native-base';
 
 import {deleteSite, updateSite} from 'terraso-client-shared/site/siteSlice';
 
-import DeleteButton from 'terraso-mobile-client/components/icons/DeleteButton';
+import DeleteButton from 'terraso-mobile-client/components/buttons/DeleteButton';
 import {TextInput} from 'terraso-mobile-client/components/inputs/TextInput';
 import {ConfirmModal} from 'terraso-mobile-client/components/modals/ConfirmModal';
 import {
